pyproject.toml
[tool.poetry]
name = "mahjong-soul-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mahjong_soul_api"}]

[tool.poetry.dependencies]
python = "^3.11"
protobuf = "4.22.3"
websockets = "11.0.2"
aiohttp = "3.8.5"
twine = "^4.0.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
